reorganize files, and fix input in footer
