+++ src/components/auth/signin-form.tsx
@@ -4,7 +4,7 @@ import { Spinner } from "@/components/loaders/spinner";
 import { GradButton } from "@/components/ui/grad-button";
 import { Input } from "@/components/ui/input";
 import { signinUser } from "@/lib/User/actions";
-import { signInSchema } from "@/lib/User/validations";
+import { signInSchema } from "@/lib/User/validations/auth";
 import { zodResolver } from "@hookform/resolvers/zod";
 import Link from "next/link";
 import { useRef } from "react";
